[r00017.1] Update bt.git : 999e502
git-reposnap-release_tag: r00017.1
git-reposnap-name: platform/system/bt
git-reposnap-prefix: system/bt
git-reposnap-fetch_url: https://source.codeaurora.org//platform/system/bt.git
git-reposnap-upstream: bt-fluoride.le.2.0.r8-rel
git-reposnap-rev: 999e502fc49320c655bb2614b09ad4b30b454e1e

commit 999e502fc49320c655bb2614b09ad4b30b454e1e
Merge: a742ee6 3d046be
Author: Linux Build Service Account <lnxbuild@localhost>
Date:   Sun Jan 28 21:43:33 2018 -0800

    Merge 3d046be32e0016477c53c1b85b1780420e7a45d6 on remote branch

+++ system/bt/stack/include/avdt_api.h
@@ -427,7 +427,7 @@ typedef struct {
 #define AVDT_DATA_OPT_NONE      0x00         /* No option still add RTP header */
 #define AVDT_DATA_OPT_NO_RTP   (0x01 << 0)   /* Skip adding RTP header */
 
-typedef UINT8 tAVDT_DATA_OPT_MASK;
+typedef UINT16 tAVDT_DATA_OPT_MASK;
